Security

+++ src/main/java/eu/ensup/my_resto/service/ItemService.java
@@ -5,8 +5,14 @@ import eu.ensup.my_resto.domain.Item;
 import eu.ensup.my_resto.model.ItemDTO;
 import eu.ensup.my_resto.repos.ImageRepository;
 import eu.ensup.my_resto.repos.ItemRepository;
+
 import java.util.List;
+import java.util.Optional;
+import java.util.Random;
 import java.util.stream.Collectors;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -15,13 +21,18 @@ import org.springframework.web.server.ResponseStatusException;
 @Service
 public class ItemService {
 
-    private final ItemRepository itemRepository;
-    private final ImageRepository imageRepository;
+    @Autowired
+    private ItemRepository itemRepository;
+    @Autowired
+    private ImageRepository imageRepository;
+    @Autowired
+    private FileService fileService;
 
-    public ItemService(final ItemRepository itemRepository, final ImageRepository imageRepository) {
-        this.itemRepository = itemRepository;
-        this.imageRepository = imageRepository;
-    }
+    @Value("${upload.path}")
+    public String uploadPath;
+
+    Random random = new Random();
+
 
     public List<ItemDTO> findAll() {
         return itemRepository.findAll()
@@ -39,6 +50,13 @@ public class ItemService {
     public Long create(final ItemDTO itemDTO) {
         final Item item = new Item();
         mapToEntity(itemDTO, item);
+        if (item.getImage() != null) {
+            String filename = randomString();
+            fileService.saveImage(item.getImage(), filename);
+            item.getImage().setPath(String.format("%s/%s", uploadPath, filename));
+            imageRepository.save(item.getImage());
+        }
+
         return itemRepository.save(item).getId();
     }
 
@@ -50,7 +68,11 @@ public class ItemService {
     }
 
     public void delete(final Long id) {
-        itemRepository.deleteById(id);
+        Optional<Item> item = itemRepository.findById(id);
+        if (item.isPresent()) {
+            itemRepository.deleteById(id);
+            fileService.deleteImage(item.get().getImage());
+        }
     }
 
     private ItemDTO mapToDTO(final Item item, final ItemDTO itemDTO) {
@@ -59,7 +81,8 @@ public class ItemService {
         itemDTO.setQuantity(item.getQuantity());
         itemDTO.setDescription(item.getDescription());
         itemDTO.setPrice(item.getPrice());
-        itemDTO.setImage(item.getImage() == null ? null : item.getImage().getId());
+        itemDTO.setCategory(item.getCategory());
+        itemDTO.setImage(item.getImage());
         return itemDTO;
     }
 
@@ -67,13 +90,22 @@ public class ItemService {
         item.setName(itemDTO.getName());
         item.setQuantity(itemDTO.getQuantity());
         item.setDescription(itemDTO.getDescription());
+        item.setCategory(itemDTO.getCategory());
         item.setPrice(itemDTO.getPrice());
-        if (itemDTO.getImage() != null && (item.getImage() == null || !item.getImage().getId().equals(itemDTO.getImage()))) {
-            final Image image = imageRepository.findById(itemDTO.getImage())
-                    .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "image not found"));
-            item.setImage(image);
-        }
+        item.setImage(itemDTO.getImage());
         return item;
+    }
+
+    public String randomString() {
+        int leftLimit = 48; // numeral '0'
+        int rightLimit = 122; // letter 'z'
+        int targetStringLength = 15;
+        String generatedString = random.ints(leftLimit, rightLimit + 1)
+                .filter(i -> (i <= 57 || i >= 65) && (i <= 90 || i >= 97))
+                .limit(targetStringLength)
+                .collect(StringBuilder::new, StringBuilder::appendCodePoint, StringBuilder::append)
+                .toString();
+        return String.format("%s.png", generatedString);
     }
 
 }
